+++ internal/perms/acl_test.go
@@ -4,6 +4,7 @@
 package perms
 
 import (
+	"context"
 	"encoding/json"
 	"fmt"
 	"testing"
@@ -25,6 +26,8 @@ type scopeGrant struct {
 func Test_ACLAllowed(t *testing.T) {
 	t.Parallel()
 
+	ctx := context.Background()
+
 	type actionAuthorized struct {
 		action       action.Type
 		authorized   bool
@@ -44,8 +47,8 @@ func Test_ACLAllowed(t *testing.T) {
 		{
 			scope: "o_a",
 			grants: []string{
-				"id=ampw_bar;actions=read,update",
-				"id=ampw_baz;actions=read:self,update",
+				"ids=ampw_bar,ampw_baz;actions=read,update",
+				"id=ampw_bop;actions=read:self,update",
 				"type=host-catalog;actions=create",
 				"type=target;actions=list",
 				"id=*;type=host-set;actions=list,create",
@@ -55,7 +58,7 @@ func Test_ACLAllowed(t *testing.T) {
 			scope: "o_b",
 			grants: []string{
 				"id=*;type=host-set;actions=list,create",
-				"id=hcst_mypin;type=host;actions=*;output_fields=name,description",
+				"ids=hcst_mypin;type=host;actions=*;output_fields=name,description",
 				"id=*;type=*;actions=authenticate",
 				"id=*;type=*;output_fields=id",
 			},
@@ -118,8 +121,18 @@ func Test_ACLAllowed(t *testing.T) {
 			},
 		},
 		{
-			name:        "matching scope and id and matching action",
+			name:        "matching scope and id and matching action first id",
 			resource:    Resource{ScopeId: "o_a", Id: "ampw_bar"},
+			scopeGrants: commonGrants,
+			actionsAuthorized: []actionAuthorized{
+				{action: action.Read, authorized: true},
+				{action: action.Update, authorized: true},
+				{action: action.Delete},
+			},
+		},
+		{
+			name:        "matching scope and id and matching action second id",
+			resource:    Resource{ScopeId: "o_a", Id: "ampw_baz"},
 			scopeGrants: commonGrants,
 			actionsAuthorized: []actionAuthorized{
 				{action: action.Read, authorized: true},
@@ -187,8 +200,18 @@ func Test_ACLAllowed(t *testing.T) {
 			},
 		},
 		{
-			name:        "matching scope, type, action, random id and bad pin",
+			name:        "matching scope, type, action, random id and bad pin first id",
 			resource:    Resource{ScopeId: "o_a", Id: "anything", Type: resource.HostCatalog, Pin: "ampw_bar"},
+			scopeGrants: commonGrants,
+			actionsAuthorized: []actionAuthorized{
+				{action: action.Update},
+				{action: action.Delete},
+				{action: action.Read},
+			},
+		},
+		{
+			name:        "matching scope, type, action, random id and bad pin second id",
+			resource:    Resource{ScopeId: "o_a", Id: "anything", Type: resource.HostCatalog, Pin: "ampw_baz"},
 			scopeGrants: commonGrants,
 			actionsAuthorized: []actionAuthorized{
 				{action: action.Update},
@@ -306,7 +329,7 @@ func Test_ACLAllowed(t *testing.T) {
 			},
 		},
 		{
-			name:        "read self with top level read",
+			name:        "read self with top level read first id",
 			resource:    Resource{ScopeId: "o_a", Id: "ampw_bar"},
 			scopeGrants: commonGrants,
 			actionsAuthorized: []actionAuthorized{
@@ -315,8 +338,17 @@ func Test_ACLAllowed(t *testing.T) {
 			},
 		},
 		{
-			name:        "read self only",
+			name:        "read self with top level read second id",
 			resource:    Resource{ScopeId: "o_a", Id: "ampw_baz"},
+			scopeGrants: commonGrants,
+			actionsAuthorized: []actionAuthorized{
+				{action: action.Read, authorized: true},
+				{action: action.ReadSelf, authorized: true},
+			},
+		},
+		{
+			name:        "read self only",
+			resource:    Resource{ScopeId: "o_a", Id: "ampw_bop"},
 			scopeGrants: commonGrants,
 			actionsAuthorized: []actionAuthorized{
 				{action: action.Read},
@@ -360,7 +392,7 @@ func Test_ACLAllowed(t *testing.T) {
 			var grants []Grant
 			for _, sg := range test.scopeGrants {
 				for _, g := range sg.grants {
-					grant, err := Parse(sg.scope, g, WithAccountId(test.accountId), WithUserId(test.userId))
+					grant, err := Parse(ctx, sg.scope, g, WithAccountId(test.accountId), WithUserId(test.userId))
 					require.NoError(t, err)
 					grants = append(grants, grant)
 				}
@@ -381,6 +413,10 @@ func Test_ACLAllowed(t *testing.T) {
 }
 
 func TestACL_ListPermissions(t *testing.T) {
+	t.Parallel()
+
+	ctx := context.Background()
+
 	tests := []struct {
 		name                        string
 		userId                      string
@@ -498,8 +534,7 @@ func TestACL_ListPermissions(t *testing.T) {
 					scope: "o_1",
 					grants: []string{
 						"id=s_1;type=session;actions=list,read",
-						"id=s_2;type=session;actions=list,read",
-						"id=s_3;type=session;actions=list,read",
+						"ids=s_2,s_3;type=session;actions=list,read",
 					},
 				},
 			},
@@ -581,8 +616,7 @@ func TestACL_ListPermissions(t *testing.T) {
 					scope: "o_1",
 					grants: []string{
 						"id=s_1;type=session;actions=list,no-op",
-						"id=s_2;type=session;actions=list,no-op",
-						"id=s_3;type=session;actions=list,no-op",
+						"ids=s_2,s_3;type=session;actions=list,no-op",
 					},
 				},
 			},
@@ -698,8 +732,7 @@ func TestACL_ListPermissions(t *testing.T) {
 					scope: "o_1",
 					grants: []string{
 						"id=s_1;type=session;actions=list,read",
-						"id=s_2;type=session;actions=list,read",
-						"id=s_3;type=session;actions=list,read",
+						"ids=s_2,s_3;type=session;actions=list,read",
 					},
 				},
 				{
@@ -815,7 +848,7 @@ func TestACL_ListPermissions(t *testing.T) {
 			var grants []Grant
 			for _, sg := range tt.aclGrants {
 				for _, g := range sg.grants {
-					grant, err := Parse(sg.scope, g, WithSkipFinalValidation(tt.skipGrantValidationChecking))
+					grant, err := Parse(ctx, sg.scope, g, WithSkipFinalValidation(tt.skipGrantValidationChecking))
 					require.NoError(t, err)
 					grants = append(grants, grant)
 				}
@@ -847,6 +880,8 @@ func TestJsonMarshal(t *testing.T) {
 func Test_AnonRestrictions(t *testing.T) {
 	t.Parallel()
 
+	ctx := context.Background()
+
 	type input struct {
 		name              string
 		grant             string
@@ -859,14 +894,28 @@ func Test_AnonRestrictions(t *testing.T) {
 			grant: "id=foobar;actions=%s",
 		},
 		{
+			name:  "ids-specific",
+			grant: "ids=foobar;actions=%s",
+		},
+		{
 			name:              "wildcard-id",
 			grant:             "id=*;type=%s;actions=%s",
 			templatedType:     true,
 			shouldHaveSuccess: true,
 		},
 		{
+			name:              "wildcard-ids",
+			grant:             "ids=*;type=%s;actions=%s",
+			templatedType:     true,
+			shouldHaveSuccess: true,
+		},
+		{
 			name:  "wildcard-id-and-type",
 			grant: "id=*;type=*;actions=%s",
+		},
+		{
+			name:  "wildcard-ids-and-type",
+			grant: "ids=*;type=*;actions=%s",
 		},
 		{
 			name:              "no-id",
@@ -883,10 +932,19 @@ func Test_AnonRestrictions(t *testing.T) {
 				if i == resource.Controller || i == resource.Worker {
 					continue
 				}
-				for j := action.Type(1); j <= action.ReadCertificateAuthority; j++ {
+				for j := action.Type(1); j <= action.Download; j++ {
+					id := "foobar"
+					prefixes := globals.ResourcePrefixesFromType(resource.Type(i))
+					if len(prefixes) > 0 {
+						id = fmt.Sprintf("%s_%s", prefixes[0], id)
+						// If it's global scope, correct it
+						if id == "global_foobar" {
+							id = "global"
+						}
+					}
 					res := Resource{
 						ScopeId: scope.Global.String(),
-						Id:      "foobar",
+						Id:      id,
 						Type:    resource.Type(i),
 					}
 					grant := test.grant
@@ -896,7 +954,7 @@ func Test_AnonRestrictions(t *testing.T) {
 						grant = fmt.Sprintf(grant, action.Type(j).String())
 					}
 
-					parsedGrant, err := Parse(scope.Global.String(), grant, WithSkipFinalValidation(true))
+					parsedGrant, err := Parse(ctx, scope.Global.String(), grant, WithSkipFinalValidation(true))
 					require.NoError(err)
 
 					acl := NewACL(parsedGrant)