+++ version/feature_manager.go
@@ -21,6 +21,7 @@ const (
 	UseTargetIdForHostId
 	RequireVersionInWorkerInfo
 	SshSessionRecording
+	SupportIdInGrants
 )
 
 var featureMap map[Feature]MetadataConstraint
@@ -71,6 +72,12 @@ func init() {
 	}
 	featureMap[SshSessionRecording] = MetadataConstraint{
 		Constraints: mustNewConstraints(">= 0.13.0"),
+	}
+
+	// Warn until 0.15 about using the now-deprecated id field in grants; after
+	// that disallow it
+	featureMap[SupportIdInGrants] = MetadataConstraint{
+		Constraints: mustNewConstraints("< 0.15.0"),
 	}
 }